Don't use forceActiveFocus to set search field focus
Instead just set the right focus properties so that we are actually
declaring what we want instead of bypassing all the focus mechanisms.

This also uses Plasmoid.expanded to ensure we reevalute the focus
binding when the applet is reshown so it is in a known state.

(cherry picked from commit 1f92caf39eab82ca09283d5be0ad250e8a3ca8ab)