Fix not working applet with same compact/full representation when hiding
If preferred, compact and full represenation are set to the same value,
the value of appletShouldBeExpanded will be true, which means the applet
will be expanded by default. In this case, only fullRepresentation will
be created and compactRepresentationItem will simply be null.

If compactRepresentationItem is null, also check the fullRepresentation
for mouse event forwarding.

This fixes manage-inputmethod applet not working if it's hidden in the
tray.