Give Breeze a Desktop Layout file
This is a follow-up Merge Request to https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/1043 to give Breeze a Desktop Layout file... so that it actually gets identified as having one.

It's just a simple script that adds Breeze's Panel and default Desktop Background engine... which results in the Vanilla KDE Plasma layout and settings you know and love when used.

This merge request does not need the MR it is a follow-up of, however it's redundant without that MR being merged.

<h1>Screenshots</h1>
<h2>Desktop Layout on its own, with appearance settings applied from Graphite Dark's Global Theme</h2>

![1](/uploads/d28c7976fc3738e7bcb0e6ad58f5e82c/1.png)

<h2>Full Breeze (Appearance + Layout)</h2>

![2](/uploads/553be47053c0ad1b18660715032018d2/2.png)