SystemDialog: give minimumWidth a cap
having the width entirely dependent on the implicit size of the content
has the unfortunate side effect that a window can become insanely wide.
instead cap the minimum width at a percentage of the screen width.
shrinking beyond that cap then only allows dimension changes in height,
not width. this allows the content items to adjust accordingly (e.g.
labels would be forced to wrap).

NB: height is intentionally uncapped because we need somewhere to put
the "overflow" content