libtaskmanager: Fix dragging launcher icon when separateLaunchers is false
When separateLaunchers is false and there are two opened pinned tasks
and one launcher icon, dragging the launcher icon to a position between
the two pinned tasks always fails for the first time.

The general idea of this fix is to make sure after every move operation,
pinned launhcer item and the corresponding opened window still stay
together.

BUG: 448912
FIXED-IN: 5.26