Klipper: Introduce a "Popup" config page with options for the popup menu
With the check box options previously on the "Actions" page, plus the
"Advanced" button (renamed to "Excluded Windows" to better reflect its
function), a shadow of the enable option (as on the system tray
popup menu) and the popup time (originally on the "General" page).
This brings all of the popup menu configuration options together,
and allows more space for the list on the "Actions" page.

Consistent wording referring to the menu throughout.

Add a message to point the user towards the next page for
configuring the actions.

Add some explanation text to the "Exclusions" list dialogue to explain
its function.