applets/systemtray: Move icon logic to the applet
This moves icon visualization logic to the qml side. The applet is
responsible for deciding whether to use IconName or IconPixmap. It also
opens the opportunity for improving visuals of overlay icons, for
example the applet could provide a cutout, etc.

In order to support the IconThemePath, the system tray model provides
the KIconLoader object, which can be passed to PlasmaCore.IconItem.

Co-authored: David Redondo <kde@david-redondo.de>