Add an explicit unlock button for passwordless users
In the refactor we correctly made it so passwordless users were not
prompted for passwords.

This is in the right direction but the UX was still slightly off as it
meant wiggling the mouse unlocked the screen, not an explicit action.

This patch adds an explicit button that must be pressed if no other
prompts were given during the authentication process.

BUG: 440055