Change UI string for "Tablet Mode" to "Touch Mode"
"Tablet Mode" is jargon that can confuse users. In this mode, things
become larger as an optimization for touch usage, but the word "tablet"
could easily refer to a drawing tablet, for which no touch
optimizations make sense.

Let's change the UI text here to call it "Touch Mode" instead.