Klipper: Tidy up the "Actions" configuration page and editing
This is the remaining part of the Klipper configuration dialogue,
the "Actions Configuration" page.  Now that the options have been
moved to the "Popup" page, this is a simple tree view with action
buttons.  The explanations and help links have been moved to the
"Edit Action" page, because that is where they are most applicable.

The "Action Properties" dialogue uses a QFormLayout, and includes
explanation text for the "Automatic" check box and the regexp help link.

Instead of editing in place, the "Edit" button opens a further dialogue
to edit the current entry.  This is more discoverable and should be easier
on small displays than editing in place, although a double click is still
accepted.  This dialogue has explanation text for substitutions.

The icon for a command can be set explicitly, for those cases where the
automatic detection (from the first word of the command) does not work.

Confirmation is requested when deleting a command or an action.

GUI:
I18N: