Revert "[sddm-theme] Clean up footer buttons code"
This partially reverts commit d2bb480706f8fec45b6941be1f7cf037c2652c30.

Something unhealthy is going on with that lastIndex that prevents it
from functioning normally when currentIndex is bound to it from the
beginning. Maybe it just needs an on*Changed() signal, so that the
instantiator has already created its objects by that time? Session button
is different from keyboard layouts button which uses pre-baked QList as
its model.