[System Tray] Emit pressed signal on compact representation in popup, too
Some actions, such as middle click to mute, trigger only on press. Some applets
also need a pressed-click because of some legacy autoclose on focus change
bug.

Unfortunately, QML does not disambiguate the "pressed" property and "pressed(mouse)"
signal with argument, so this has to be done awkwardly through QMetaObject on C++ side.

BUG: 426646
BUG: 452893
FIXED-IN: 5.25.0