Lock screen: Avoid wiping password field when getting our first prompt
When the lockscreen is in the screensaver mode we want the keyboard key
pressed to wake the screen to go to the password box textfield. This did
work correctly, but a code path also reset it when we get the first
prompt.

In the (currently hypothetical) case of multiple prompts we would want
to clear anything in the prompt.

This uses the existing boolean flag to handle that appropriately.

BUG: 454108


(cherry picked from commit 74a80aaac78647c51662ca58e43078ef681059c3)