Set up session for proper Xwayland clients scaling
If KWin Wayland is set up to let Xwayland clients to scale themselves,
set the configuration values we need for them to do so properly.
To do so, we extend the fonts kcminit module to provide a meaningful
default DPI to scale to and we set some environment variables for
applications to scale as expected. These are the ones we already were
using to properly scale in X11 sessions.