NoPaswordUnlock screen: focus it by default
Otherwise the user needs to manually tab over to the Unlock button
first, which is annoying. And they may not even think to do it at all.

The `forceActiveFocus()` is because the stack view does not manually pass focus
to the item you push onto it; this needs to be handled manually.

BUG: 454707


(cherry picked from commit 370976d71582c4d057147dd6f123acc4bcf5d010)