startkde: Reload systemd on Plasma start
If the time between logout and relogin is short enough or a separate session
is open during that time, the systemd user instance keeps running. This means
that generators like the one for XDG autostart handling aren't run again, which
effectively means it's still using the configuration from the previous session.

Just reload systemd on every login to make sure its state is fresh.