[applets/batterymonitor] Restructure GridLayout based table into model-view
Time to change my mind. I was the one who earlier refactored it from an
(abusive) model based implementation to a simply bunch of components.
But now I found the perfect way to reparent whole rows of Items to a
GridLayout -- or any other layout type for that matter. The secret, as
always in QML land, is to have some deeply magical very built-in and
insanely integrated way of doing things: extra Repeater + ObjectModel.