[lookandfeel] Fix (workaround) QML Binding on list<> bug
When I tried to change targets dynamically (like, to animate only
current user item in the list view), turned out that Binding fails to
restore the transform property back to an empty list, probably because
a list is not a "value" in QML engine's blind eyes, nor does it have
any WRITE or NOTIFY members in QQuickItem class. And we can't even hook
into the Binding::targetChanged since there's not targetChanged signal
to begin with! Most Binding's properties are just not notifiable.

This manual imperative approach also breaks any preexisting bindings as
before, but we don't really care about restoring such a marginal and
rare property, at least not in this scenario where all the targets are
known in advance.