[lookandfeel] Fix rejection animations: only animate specific items
This is self-explanatory. We received some amount of complaints about
too much motion on the screen. In case of users list this is obviously
and objectively true -- I just didn't test on a wide list with lots of
users; with action buttons row it is more opinionated but that's fine.