Fix password field in lock screen not clearing after failed login attempt
This bug was introduced by plasma/plasma-workspace!1754.

`lockScreenUi.hadPrompt` was set to `false` in `onFailed()` to avoid
unintentionally clearing the password after the text field had faded
out after 10 seconds and the user pressed a button. Unfortunately this
also caused the password field to not clear right after the failed event.

So instead the statement is moved into the trigger of `fadeoutTimer` which
fixes both issues.

BUG: 455227
FIXED-IN: 5.25.3