wallpapers/image: fix default image overriding the true image
The default image is set in MediaProxy ctor. When the default image is
still in loading status, the model image can change and be already
loaded before the default image is loaded, therefore there could be a
race condition that the default image is loaded later and overrides the
true image.

This stores the pending image and destroys the previous pending image
to fix the race condition.

BUG: 456189
FIXED-IN: 5.26