Take active screen in account and remove shortcut requirement for activating launcher
With the information about the active screen we can make an educated
guess about where the attention of the user is and where they could
expect the launcher to open. If on the screen no launcher could be
activated, look at all launchers. Also remove the requirement for
launchers to have a shortcut. This is needed to make this feature
work reliably and should reduce the instances of "Meta key stopped working"
happening in general.
CCBUG:447962
BUG:391322
BUG:444343
BUG:437979