lockscreen KCM: clarify clock and media control UI
The current UI does not communicate to the user at what times these UI
elements will be shown, leading to confusion. People are thinking that
unchecking the Clock checkbox always hides it, but this is not the
case; instead it hides the clock only when the main UI is invisible.

To address this, the wordings of the checkbox labels are changed to
indicate what the opposite states will do as well.

BUG: 429468
FIXED-IN: 5.26