kcms/lookandfeel: Allow themes to set titlebar button layout and BorderlessMaximizedWindows
This merge request adds support for additional Global Theme layout settings to
be set in KCM, if the selected theme supports them:

- (controlled by Desktop Layout switch) Borderless Maximized Windows
- (own switch) Titlebar Buttons Layout

Ones with their own switch also have new checkboxes for themselves in More Options in
the KCM, and the values for all the added options, in the new `layouts/defaults` file
of Global Themes, are made to match their KDE Plasma config files counterparts in names
and values.

Example of a `layouts/defaults` file in a Global Theme:

```
[kwinrc][org.kde.kdecoration2]
ButtonsOnLeft=M
ButtonsOnRight=IAX

[kwinrc][Windows]
BorderlessMaximizedWindows=false
```

NOTE: If a Global Theme doesn't supply `BorderlessMaximizedWindows`, it'll be turned off automatically when applying a Desktop Layout to prevent issues.