[lockscreen] Fix lockscreen not unlocking or showing extra "Unlock" button
Turns out aliasing a property that is inside a component which is
declared inside another component that is bound to a default property
which in turn itself is an alias to some item's children it just too
must for QML engine: it completely messes up such property's metadata
so it could be pointing to any random memory location.

Surprisingly, an alias to an object (the one owning that property) still
works fine, so as a workaround let's just remove the showPassword alias,
and set it via `mainPasswordBox`.

See also: QTBUG-xxxxx

BUG: 456210