kcms/colors: port away from ugly hacks in saveColors
Use the proper DBus signal to notify the change of
accentColorFromWallpaper.

applyScheme needs the latest accentColor but when it's called from `KCMColors::saveColors()`,
the accent color in kdeglobals is still old, that's why the hack works
by saving the value before `ManagedConfigModule::save()`. Therefore we
can pass the latest accent color value to applyScheme to fix it and
remove the ugly hack.