kcms/users: don't let "Choose File…" text overflow
When using a language that's wordier than English and this label becomes
a two-line string, currently the layout breaks and the string overflows
because the Label item that draws it has no maximum height and elision
mode set.

This commit fixes that by setting them accordingly, and also using a
standard icon size for the icon so that there's actually enough room for
two lines of text.

BUG: 458614
FIXED-IN: 5.24.7
(cherry picked from commit 0dad5f0789c8eb5e4fa893622fdc550898488779)