kcms/nightcolor: hide preview message after scroll manipulation
Currently, if you interact with one of the color temperature sliders
by scrolling, the preview message and full-screen color tint are
activated but never deactivated. This is because the deactivation only
happens in an `onPressedChanged` handler, but with a scroll, there is no
change to the `pressed` property's status, so it never fires and hence
the message and overlay never disappear.

This commit fixes that by using a timer to hide them that only triggers
when a slider is manipulated using a scroll.

BUG: 458675
FIXED-IN: 5.26