containmentactions/contextmenu: Add "Open Terminal" action
This adds the ability to open the terminal from the main Plasma Shell
context menu. It was originally written as a Red Hat/Fedora patch
for KDE Workspace 4 to open Konsole from the desktop context menu
when Fedora switched to KDE 4, and ported forward to KDE Plasma 5
when Fedora Linux switched to Plasma 5 in Fedora 22.

It has been maintained as a downstream patch since then.

This version of the patch genericizes the patch by not implying that
Konsole is the user's terminal (it actually opens the user's chosen
terminal anyway), can be restricted via the desktop kiosk mode
settings, and is disabled by default.

The major maintainers of this patch are being honored as co-authors
for this patch. They have all helped keep this functionality working
across two KDE major versions for a decade.

Co-authored-by: Daniel Vrátil <dvratil@redhat.com>
Co-authored-by: Jan Grulich <jgrulich@redhat.com>
Co-authored-by: Marc Deop <marcdeop@fedoraproject.org>
Co-authored-by: Rex Dieter <rdieter@fedoraproject.org>

BUG: 451217