apples/kicker: fix timer in triangular mouse filter.
The underlying mouse area for kickoff list delegates does not select
items when the mouse enters the hovering area, but when the mouse
changes its position.
Therefore, to change the current selected item, the triangular mouse
filter must replay both enter and move events when the triangular
fitlering condition is broken.

Sometimes, it was possible that the filter only let trough an enter
and no following move event, that results in the current hovered item
not selected even when the mouse stops its movement completely.

The timeout timer that is used to select the hovered item after a time
was not at fault, because it is only responsible to do so when the
triangular filtering condition is currently upheld, which is not the
case in the above mentioned example.

BUG: 438950