wallpapers/image: set duration to 1 to avoid flickering
When Image is in a Rectangle, though the image is set to fill the
rectangle, the rectangle can still be seen before the image is fully
loaded.

If asynchronous is set to true, the exact moment Image.status becomes
Image.Ready, the image doesn't really show up in the window.

To verify the behavior, run qml main.qml in the attachment and click
the window area (will switch between the two images), see if every
time the color of rectangle shows up before the image.

See also: https://bugreports.qt.io/browse/QTBUG-106797