[SDDM] Don't use margins for user avatar
Currently, the SDDM uses the default picture for a user who doesn't have an avatar, while the lock/logout screens in this case use a specific system icon. These picture and icon are the same (but located in different places), but since the margins are set for the icon, they look different on the sddm and the lock/logout screens. This MR fixes that.