[applet/appmenu] Don't reset menu when popups become focused to avoid extra flickering
Test plan:

a) receive a notification while an app with a menu is active, select
text on the notification twice (there seems to be a bug in task manager
that it only updates currently active window on a second click).

Expected result: menu should stay.

b) Put a Global Menu widget on a desktop, open an app which has a menu.
Click on a menu.

Expected result: now it is possible to interact with it without a menu
instantly disappearing.

c) Open an app which has a menu on one virtual desktop, then switch to
another virtual desktop which doesn't have any windows open (so that
the Plasma desktop becomes focused).

Excepted result: menu should disappear.

d) Open some apps with menus on two different activities. Focus them on
both activities. Switch between activities via shortcut (not via pager
which might take the focus away).

Expected result: menu updates correctly.

e) Open an app with a menu on one activity, leave other activity empty.
Focus that app, and switch via shortcut as in previous test case.

Expected result: menu hides on an empty activity, but comes back on the
other one.

f) Open a single window with a menu on an otherwise empty desktop.
Minimize the window.

Expected result: menu hides when the window is minimized, and only
the desktop remains.

BUG: 404857
CCBUG: 444587
CCBUG: 449781
BUG: 449251