[applet/appmenu] Guard the menu against late unintended updates
Problem:

When virtual desktops or activities are switched, app menu gets a batch
or back-to-back updates. For example, when switching to an activity with
no windows, an applet first gets a signal with the same app id and sends
an update request, then immediately arrives the second signal without
valid task index, so an applet hides the menu, but after that some time
later it will receive a reply from D-Bus and will update the menu even
though it is not needed anymore.

Solution: let's guard the reply handler with a simple boolean flag.

It doesn't fix much now, but will become important in the following patch.