System Tray: Make Media Player and Notifications applets SystemServices
Right now these are marked as ApplicationStatus, which means they appear
in the System Tray in the same logical group as app icons from the likes
of Steam, Discord, Telegram etc. This is sort of weird since those icons
are app status notifiers (SNIs) and focus an app when clicked, while the
Media Player and Notifications applets are applets that show a popup when
clicked. We have gotten a few user complaints about it feeling weird to
have them interspersed.

And the Media Player and Notifications applets do show application
status, but they don't show status for a single application, which
makes them more like system services. And in fact, they are system
services--just system services that get their content from applications.
So moving them to the System Services category isn't semantically
incorrect. Let's do it to improve the grouping here.

CCBUG: 362077