Remove ContainmentType from the systray container plasmoid
The systray container plasmoid is *not* a containment, but is a normal
applet that does some hack to display the contents of the internal
systray applet which *is* one.

Since plasma-framework is built with
if KCOREADDONS_BUILD_DEPRECATED_SINCE(5, 89)

now everything that has the containmentype key is considered as a
containment, and now the systray is listed in the "add panel" menu,
however it doesn't really work there.
that key was wrong from there to begin wwith. After some testing
there doesn't seem to be any behavior difference of the systary plasmoid
when used

BUG: 454416
FIXED-IN: 5.26