components/calendar + applets/digital-clock: Remove partial fork in digital-clock
The digital clock uses a slightly different layout in the header and has an
unnecessary fork of header from calendar. This presents a code maintenance issue
as the header and calendar body are deeply related, and changes to calendar often
break the digital clock and vice-versa. Controlling them in one place makes them
maintainable and more pleasant to work with.