kcms/colors: Fix color previews not updating until apply is pressed
The reason colors are not updating correctly because there is no color(valid) to show in the first place. The color generation was later disabled when setting is not from wallpaper after the feature being implemented, I guess for efficiency. So we need to temporarily turn on the color engine, take the color and then again disable it. There is two reason for which we should always return a valid color for primary desktop:
1. We need to show the preview when asked.
2. There is a public dbus api which was added so that any third party code can get the correct wallpaper accent color any moment, regardless of setting. This may be useful, for example some app which always want to show some ui element in the wallpaper color.

BUG: 456835
FIXED-IN: 5.26


(cherry picked from commit 5b7c5a201dc2517a5545a7017caf10bf78dde335)