BasicAppletContainer: Refactor blurEnabled code
Turning it into a binding allows receiving updates when GraphicsInfo API
changes, and now it also includes a dependency on effectiveBackgroundHints.

Overall, this patch frees opacity mask (background blur) from a
dependency on appletContainer.applet object, so it will be easier to
factor the whole mask out as a component.