Surface middle-clickability in the UI for middle-clickable applets
Various System Tray applets can be middle-clicked to toggle some
behavior, but this is a hidden feature you would never discover if you
didn't read Kai's blog. :) The features are quite handy, so let's expose
them to more users by mentioning them in the tooltip subtext for the
relevant applets.