[kcms/icons] Use OverlaySheet for icon size popup
Right now it's a custom pop-up which does basically the same thing as a
Kirigami OverlaySheet, but it does not use that component, so as a result
it is visually inconsistent.

This commit ports it to use Kirigami.OverlaySheet, making whatever
minor layout changes to the UI are necessary to make it fit well inside
the sheet.

Everything still works including focus and arrow key handling.

cc @broulik @teams/vdg 

Here's how it looks now:

![image](/uploads/8fea8afdd5e2ddaf732891854c86d2be/image.png)