wallpapers/image: fall back to default wallpaper when url is empty
It's expected to use the default wallpapaer when url is empty, which is
also the default value.

When closing the wallpaper config dialog without saving config, the
source property of MediaProxy will become empty if the previous
wallpaper is the default wallpaper, but in setSource empty url will
be directly rejected instead of switching to the default wallpaper.

This fixes the wrong behavior by calling `useSingleImageDefaults` when
url is empty, and reset the source url in `useSingleImageDefaults` so it
can find a wallpaper from the current Plasma theme.

BUG: 460692


(cherry picked from commit 0eb8db0e70330bffc3abfb6c51859e6ab1a3cb70)