SystemDialog: stop using two data sources for titleHeading label
Right now the titleHeading label gets its text from two places:
1. its `text` property is directly bound to root's `title` property
2. its `text` property is also aliased to root's `mainText` property

This creates a non-deterministic multiple-sources-of-truth problem and
also causes an odd UX whereby if only the title property is defined,
the same text is shown in both the window's titlebar, and also right
below it in the window content itself. This is redundant.

To fix this, let's only get the text from the mainText property and let
the title property only set the window title.